rename use-new-tag
